rearrange into cryptography.py
